requirements.txt
openai
moviepy
glob
urllib
json
PIL
python-dotenv
suno-api
